:fire: remove role & policy tables

+++ src/plugins/nut/services/locale.rs
@@ -1,3 +1,4 @@
+use std::any::type_name;
 use std::ops::{Deref, DerefMut};
 use std::sync::{Arc, Mutex};
 
@@ -39,7 +40,15 @@ impl v1::locale_server::Locale for Service {
         let db = db.deref_mut();
         let jwt = self.jwt.deref();
         let user = try_grpc!(ss.current_user(db, jwt))?;
-        try_grpc!(user.can::<Locale>(db, &Operation::Write, None))?;
+
+        if !has_permission!(
+            self.enforcer,
+            &user.subject(),
+            Operation::Write.to_string(),
+            to_resource!(type_name::<Locale>())
+        ) {
+            return Err(Status::permission_denied(type_name::<Locale>()));
+        }
 
         let req = req.into_inner();
         let code = to_code!(req.code);
@@ -80,7 +89,14 @@ impl v1::locale_server::Locale for Service {
         let db = db.deref_mut();
         let jwt = self.jwt.deref();
         let user = try_grpc!(ss.current_user(db, jwt))?;
-        try_grpc!(user.can::<Locale>(db, &Operation::Write, None))?;
+        if !has_permission!(
+            self.enforcer,
+            &user.subject(),
+            Operation::Remove.to_string(),
+            to_resource!(type_name::<Locale>())
+        ) {
+            return Err(Status::permission_denied(type_name::<Locale>()));
+        }
         let req = req.into_inner();
         try_grpc!(LocaleDao::delete(db, req.id))?;
         Ok(Response::new(()))
